# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nakudin Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cloudinary/signature:
    post:
      operationId: getCloudinarySignature
      tags: [cloudinary]
      summary: Get a signed upload signature for Cloudinary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloudinarySignatureRequest"
      responses:
        "200":
          description: Signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CloudinarySignature"
...
  /admin/shops:
    get:
      operationId: adminListShops
      tags: [admin]
      summary: Admin — list all shops
      parameters:
        - in: query
          name: cursor
          schema: { type: ["string", "null"] }
        - in: query
          name: status
          schema: { type: ["string", "null"] }
        - in: query
          name: q
          schema: { type: ["string", "null"] }
        - in: query
          name: limit
          schema: { type: integer, default: 30 }
      responses:
        "200":
          description: Admin shop list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminShopPage"
  /admin/shops/{shopId}:
    patch:
      operationId: adminUpdateShop
      tags: [admin]
      summary: Admin — update a shop (verify, suspend, adjust status)
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminShopUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"

    delete:
      operationId: adminDeleteShop
      tags: [admin]
      summary: Admin — delete a shop
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: Admin — list all products
      parameters:
        - in: query
          name: cursor
          schema: { type: ["string", "null"] }
        - in: query
          name: q
          schema: { type: ["string", "null"] }
        - in: query
          name: limit
          schema: { type: integer, default: 30 }
      responses:
        "200":
          description: Admin product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductPage"
  /admin/reports:
    get:
      operationId: adminListReports
      tags: [admin]
      summary: Admin — list all reports
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [open, reviewed, dismissed]
            default: open
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
  /admin/reports/{reportId}:
    patch:
      operationId: adminUpdateReport
      tags: [admin]
      summary: Admin — update report status
      parameters:
        - in: path
          name: reportId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
components:
  schemas:
...
    Product:
      type: object
      required: [id, shopId, title, price, images, likeCount, viewCount, whatsappClickCount, status, createdAt]
      properties:
        id: { type: string }
        shopId: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        images: { type: array, items: { type: string } }
        category: { type: ["string", "null"] }
        locationCity: { type: ["string", "null"] }
        locationState: { type: ["string", "null"] }
        locationLat: { type: ["number", "null"] }
        locationLng: { type: ["number", "null"] }
        likeCount: { type: integer }
        viewCount: { type: integer }
        whatsappClickCount: { type: integer }
        trendScore: { type: ["number", "null"] }
        status: { type: string, enum: [active, hidden, deleted] }
        isLiked: { type: boolean }
        createdAt: { type: string }
...
    Shop:
      type: object
      required: [id, businessName, category, followerCount, followingCount, verified, subscriptionStatus, createdAt]
      properties:
        id: { type: string }
        businessName: { type: string }
        bio: { type: ["string", "null"] }
        category: { type: string }
        logoUrl: { type: ["string", "null"] }
        coverUrl: { type: ["string", "null"] }
        whatsappNumber: { type: ["string", "null"] }
        locationCity: { type: ["string", "null"] }
        locationState: { type: ["string", "null"] }
        locationLat: { type: ["number", "null"] }
        locationLng: { type: ["number", "null"] }
        verified: { type: boolean }
        followerCount: { type: integer }
        followingCount: { type: integer }
        totalViews: { type: integer }
        totalLikes: { type: integer }
        totalWhatsappClicks: { type: integer }
        avgRating: { type: ["number", "null"] }
        reviewCount: { type: integer }
        subscriptionStatus: { type: string, enum: [trial, active, grace, locked] }
        trialEndsAt: { type: ["string", "null"] }
        isFollowed: { type: boolean }
        createdAt: { type: string }
...
    Report:
      type: object
      required: [id, targetType, targetId, reason, status, createdAt]
      properties:
        id: { type: string }
        targetType: { type: string, enum: [shop, product] }
        targetId: { type: string }
        reporterUid: { type: ["string", "null"] }
        reason: { type: string }
        details: { type: ["string", "null"] }
        status: { type: string, enum: [open, reviewed, dismissed] }
        createdAt: { type: string }
...
    ReportStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [reviewed, dismissed] }
...
    CloudinarySignatureRequest:
      type: object
      properties:
        folder: { type: string }
    CloudinarySignature:
      type: object
      required: [signature, timestamp, apiKey, cloudName, uploadPreset]
      properties:
        signature: { type: string }
        timestamp: { type: integer }
        apiKey: { type: string }
        cloudName: { type: string }
        uploadPreset: { type: string }
...
    AdminShopPage:
      type: object
      required: [shops, nextCursor]
      properties:
        shops:
          type: array
          items:
            $ref: "#/components/schemas/Shop"
        nextCursor: { type: ["string", "null"] }
    AdminProductPage:
      type: object
      required: [products, nextCursor]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        nextCursor: { type: ["string", "null"] }
    AdminShopUpdate:
      type: object
      properties:
        verified: { type: boolean }
        subscriptionStatus: { type: string, enum: [trial, active, grace, locked] }
        suspended: { type: boolean }