# lib/api-zod/src/generated/types/adminShopUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nakudin Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminShopUpdateSubscriptionStatus } from './adminShopUpdateSubscriptionStatus';

export interface AdminShopUpdate {
  verified?: boolean;
  subscriptionStatus?: AdminShopUpdateSubscriptionStatus;
  suspended?: boolean;
}