# lib/api-zod/src/generated/types/reportStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nakudin Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportStatusUpdateStatus } from './reportStatusUpdateStatus';

export interface ReportStatusUpdate {
  status: ReportStatusUpdateStatus;
}